refactor: Move ActionMenuImportDropdown from Placeholder list items
