+++ src/components/Placeholders/Placeholder.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState, Fragment, useMemo, useRef } from 'react'
+import React, { forwardRef } from 'react'
 import PropTypes from 'prop-types'
 
 import ListItem from 'cozy-ui/transpiled/react/MuiCozyTheme/ListItem'
@@ -11,41 +11,15 @@ import InfosBadge from 'cozy-ui/transpiled/react/InfosBadge'
 import FileDuotoneIcon from 'cozy-ui/transpiled/react/Icons/FileDuotone'
 import Plus from 'cozy-ui/transpiled/react/Icons/Plus'
 
-import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import { useScannerI18n } from 'src/components/Hooks/useScannerI18n'
-import { usePapersDefinitions } from 'src/components/Hooks/usePapersDefinitions'
 import { PaperDefinitionsPropTypes } from 'src/constants/PaperDefinitionsPropTypes'
-import ActionMenuImportDropdown from 'src/components/Placeholders/ActionMenuImportDropdown'
 
-const Placeholder = ({ placeholder, divider }) => {
+const Placeholder = forwardRef(({ placeholder, divider, onClick }, ref) => {
   const scannerT = useScannerI18n()
-  const actionBtnRef = useRef()
-  const { papersDefinitions } = usePapersDefinitions()
-  const [isImportDropdownDisplayed, setIsImportDropdownDisplayed] =
-    useState(false)
-  const { setCurrentDefinition } = useStepperDialog()
-  const formModel = useMemo(() => {
-    return papersDefinitions.find(
-      paper => paper.label && paper.label === placeholder.label
-    )
-  }, [papersDefinitions, placeholder.label])
-
-  const hideImportDropdown = useCallback(
-    () => setIsImportDropdownDisplayed(false),
-    []
-  )
-  const showImportDropdown = useCallback(() => {
-    if (formModel) {
-      // Set Dialog modal
-      setCurrentDefinition(formModel)
-      // Set ActionMenu
-      setIsImportDropdownDisplayed(true)
-    }
-  }, [formModel, setCurrentDefinition])
 
   return (
     <>
-      <ListItem button onClick={showImportDropdown} ref={actionBtnRef}>
+      <ListItem button onClick={() => onClick(placeholder)} ref={ref}>
         <ListItemIcon>
           <InfosBadge
             badgeContent={
@@ -76,16 +50,10 @@ const Placeholder = ({ placeholder, divider }) => {
         </Typography>
       </ListItem>
       {divider && <Divider variant="inset" component="li" />}
-
-      <ActionMenuImportDropdown
-        isOpened={isImportDropdownDisplayed}
-        placeholder={placeholder}
-        onClose={hideImportDropdown}
-        anchorElRef={actionBtnRef}
-      />
     </>
   )
-}
+})
+Placeholder.displayName = 'Placeholder'
 
 Placeholder.propTypes = {
   placeholder: PaperDefinitionsPropTypes.isRequired,
